Terminal abc (#242)
* Revert terminal check logic: FORCE_DOCKER_TERMINAL -> ALLOW_LOCAL_TERMINAL

* Enforce remote terminal requirement for BashTool and GrepTool unless local terminal is explicitly allowed

* Use `os.environ.get` instead of `os.getenv`

* Rename Terminal to LocalTerminal

* More local terminal

* Remove unused old terminal.py

* ABC Terminal and abstractmethod

* Fix imports

* Removed duplicated terminal tests

* Remove unused imports in test_terminal.py
