+++ debug_gym/gym/terminals/__init__.py
@@ -1,5 +1,6 @@
 from debug_gym.gym.terminals.docker import DockerTerminal
 from debug_gym.gym.terminals.kubernetes import KubernetesTerminal
+from debug_gym.gym.terminals.local import LocalTerminal
 from debug_gym.gym.terminals.terminal import Terminal
 from debug_gym.logger import DebugGymLogger
 
@@ -21,7 +22,7 @@ def select_terminal(
         case "kubernetes":
             terminal_class = KubernetesTerminal
         case "local":
-            terminal_class = Terminal
+            terminal_class = LocalTerminal
             if any(cfg in terminal_config for cfg in docker_only):
                 logger.warning("Ignoring Docker-only parameters for local terminal.")
         case _:
