+++ debug_gym/gym/tools/grep.py
@@ -91,13 +91,13 @@ class GrepTool(EnvironmentTool):
             # Assert that the terminal is not a local terminal (only in production)
             import os
 
-            from debug_gym.gym.terminals.terminal import Terminal
+            from debug_gym.gym.terminals.local import LocalTerminal
 
             # Require remote terminal unless local is explicitly allowed
             require_remote = (
                 os.environ.get("ALLOW_LOCAL_TERMINAL", "false").lower() == "false"
             )
-            if require_remote and type(environment.terminal) is Terminal:
+            if require_remote and type(environment.terminal) is LocalTerminal:
                 return Observation(
                     self.name,
                     "Error: grep tool requires a non-local terminal. Current terminal type is not supported.",
