+++ tests/gym/envs/test_env.py
@@ -6,10 +6,8 @@ import pytest
 
 from debug_gym.gym.entities import EvalOutput, Event, Observation
 from debug_gym.gym.envs.env import EnvInfo, EventHooks, RepoEnv, TooledEnv
-from debug_gym.gym.terminals.terminal import Terminal
 from debug_gym.gym.tools.tool import ToolCall
 from debug_gym.gym.tools.toolbox import Toolbox
-from debug_gym.gym.workspace import Workspace
 
 
 @pytest.fixture
