+++ tests/gym/terminals/test_terminal.py
@@ -1,26 +1,90 @@
-import os
-import re
 import tempfile
 from pathlib import Path
 
 import pytest
 
-from debug_gym.gym.terminals import select_terminal
+from debug_gym.gym.terminals import (
+    DockerTerminal,
+    LocalTerminal,
+    ShellSession,
+    select_terminal,
+)
 from debug_gym.gym.terminals.shell_session import DEFAULT_PS1
-from debug_gym.gym.terminals.terminal import Terminal
 
 
-def test_terminal_init():
-    terminal = Terminal()
+@pytest.fixture
+def tmp_dir_prefix(tmp_path):
+    """Expected tmp_dir_prefix for a terminal session."""
+    tmp_dir = tempfile.TemporaryDirectory(prefix="Terminal-")
+    tmp_dir_prefix = str(Path(tmp_dir.name).resolve()).split("Terminal-")[0]
+    return tmp_dir_prefix
+
+
+@pytest.if_is_linux
+def test_shell_session_run(tmp_path):
+    working_dir = str(tmp_path)
+    shell_command = "/bin/bash --noprofile --norc"
+    env_vars_1 = {"TEST_VAR": "TestVar"}
+    session_1 = ShellSession(
+        shell_command=shell_command,
+        working_dir=working_dir,
+        env_vars=env_vars_1,
+    )
+    session_2 = ShellSession(
+        shell_command=shell_command,
+        working_dir=working_dir,
+    )
+
+    assert session_1.shell_command == shell_command
+    assert session_2.shell_command == shell_command
+
+    assert session_1.working_dir == working_dir
+    assert session_2.working_dir == working_dir
+
+    assert session_1.env_vars == env_vars_1 | {"PS1": DEFAULT_PS1}
+    assert session_2.env_vars == {"PS1": DEFAULT_PS1}
+
+    output = session_1.run("echo Hello World", timeout=5)
+    assert output == "Hello World"
+
+    session_2.run("export TEST_VAR='FooBar'", timeout=5)
+    output = session_2.run("echo $TEST_VAR", timeout=5)
+    assert output == "FooBar"
+
+    output = session_1.run("echo $TEST_VAR", timeout=5)
+    assert output == "TestVar"
+
+
+def test_shell_session_timeout(tmp_path):
+    working_dir = str(tmp_path)
+    # Write a long-running command to a file
+    long_running_command = "sleep 60"
+
+    shell = ShellSession(
+        shell_command="/bin/bash --noprofile --norc",
+        working_dir=working_dir,
+    )
+
+    timeout = 1
+    with pytest.raises(
+        TimeoutError,
+        match=f"Read timeout after {timeout}",
+    ):
+        shell.run(long_running_command, timeout=timeout)
+    assert shell.is_running is False
+
+
+def test_terminal_init(tmp_dir_prefix):
+    terminal = LocalTerminal()
     assert terminal.session_commands == []
     assert terminal.env_vars["NO_COLOR"] == "1"
     assert terminal.env_vars["PS1"] == DEFAULT_PS1
     assert len(terminal.env_vars) > 2  # NO_COLOR, PS1 + os env vars
-    assert os.path.basename(terminal.working_dir).startswith("Terminal-")
+    assert terminal.working_dir.startswith(tmp_dir_prefix)
 
 
 def test_terminal_init_no_os_env_vars():
-    terminal = Terminal(include_os_env_vars=False)
+    terminal = LocalTerminal(include_os_env_vars=False)
     assert terminal.env_vars == {"NO_COLOR": "1", "PS1": DEFAULT_PS1}
 
 
@@ -28,7 +92,7 @@ def test_terminal_init_with_params(tmp_path):
     working_dir = str(tmp_path)
     session_commands = ["echo 'Hello World'"]
     env_vars = {"ENV_VAR": "value"}
-    terminal = Terminal(working_dir, session_commands, env_vars)
+    terminal = LocalTerminal(working_dir, session_commands, env_vars)
     assert terminal.working_dir == working_dir
     assert terminal.session_commands == session_commands
     assert terminal.env_vars["NO_COLOR"] == "1"
@@ -43,7 +107,7 @@ def test_terminal_init_with_params(tmp_path):
 
 def test_terminal_run(tmp_path):
     working_dir = str(tmp_path)
-    terminal = Terminal(working_dir=working_dir)
+    terminal = LocalTerminal(working_dir=working_dir)
     entrypoint = "echo 'Hello World'"
     success, output = terminal.run(entrypoint, timeout=1)
     assert success is True
@@ -51,69 +115,13 @@ def test_terminal_run(tmp_path):
     assert terminal.working_dir == working_dir
 
 
-def test_terminal_run_tmp_working_dir():
-    terminal = Terminal()
-    entrypoint = "pwd -P"
+def test_terminal_run_tmp_working_dir(tmp_path, tmp_dir_prefix):
+    terminal = LocalTerminal()
+    entrypoint = "echo 'Hello World'"
     success, output = terminal.run(entrypoint, timeout=1)
     assert success is True
-    assert output == terminal.working_dir
-
-
-@pytest.mark.parametrize(
-    "command",
-    [
-        ["echo Hello", "echo World"],
-        "echo Hello && echo World",
-    ],
-)
-def test_terminal_run_multiple_commands(tmp_path, command):
-    working_dir = str(tmp_path)
-    terminal = Terminal(working_dir=working_dir)
-    success, output = terminal.run(command, timeout=1)
-    assert success is True
-    assert output == "Hello\nWorld"
-
-
-def test_terminal_run_failure(tmp_path):
-    working_dir = str(tmp_path)
-    terminal = Terminal(working_dir=working_dir)
-    entrypoint = "ls non_existent_dir"
-    success, output = terminal.run(entrypoint, timeout=1)
-    assert success is False
-    # Linux: "ls: cannot access 'non_existent_dir': No such file or directory"
-    # MacOS: "ls: non_existent_dir: No such file or directory"
-    pattern = r"ls:.*non_existent_dir.*No such file or directory"
-    assert re.search(pattern, output)
-
-
-def test_terminal_session(tmp_path):
-    working_dir = str(tmp_path)
-    command = "echo Hello World"
-    terminal = Terminal(working_dir=working_dir)
-    assert not terminal.sessions
-
-    session = terminal.new_shell_session()
-    assert len(terminal.sessions) == 1
-    output = session.run(command, timeout=1)
     assert output == "Hello World"
-
-    session.run("export TEST_VAR='FooBar'", timeout=1)
-    output = session.run("pwd", timeout=1)
-    assert output == working_dir
-    output = session.run("echo $TEST_VAR", timeout=1)
-    assert output == "FooBar"
-
-    terminal.close_shell_session(session)
-    assert not terminal.sessions
-
-
-def test_terminal_multiple_session_commands(tmp_path):
-    working_dir = str(tmp_path)
-    session_commands = ["echo 'Hello'", "echo 'World'"]
-    terminal = Terminal(working_dir, session_commands)
-    status, output = terminal.run("pwd", timeout=1)
-    assert status
-    assert output == f"Hello\nWorld\n{working_dir}"
+    assert terminal.working_dir.startswith(tmp_dir_prefix)
 
 
 def test_select_terminal_default():
@@ -126,8 +134,16 @@ def test_select_terminal_default():
 def test_select_terminal_local():
     config = {"type": "local"}
     terminal = select_terminal(config)
-    assert isinstance(terminal, Terminal)
+    assert isinstance(terminal, LocalTerminal)
     assert config == {"type": "local"}  # config should not be modified
+
+
+@pytest.if_docker_running
+def test_select_terminal_docker():
+    config = {"type": "docker"}
+    terminal = select_terminal(config)
+    assert isinstance(terminal, DockerTerminal)
+    assert config == {"type": "docker"}  # config should not be modified
 
 
 def test_select_terminal_unknown():
@@ -141,7 +157,7 @@ def test_select_terminal_invalid_config():
 
 
 def test_shell_session_start_with_session_commands(tmp_path):
-    terminal = Terminal(
+    terminal = LocalTerminal(
         working_dir=str(tmp_path),
         session_commands=["echo setup"],
     )
@@ -172,7 +188,7 @@ def test_shell_session_start_with_session_commands(tmp_path):
 
 
 def test_shell_session_start_without_session_commands(tmp_path):
-    terminal = Terminal(working_dir=str(tmp_path))
+    terminal = LocalTerminal(working_dir=str(tmp_path))
     session = terminal.new_shell_session()
 
     # Test starting without command
@@ -197,31 +213,3 @@ def test_shell_session_start_without_session_commands(tmp_path):
     output = session.run("print('test python')", ">>>")
     assert output == "test python"
     session.close()
-
-
-def test_copy_content(tmp_path):
-    # Create a temporary source file
-    source_dir = tmp_path / "source_dir"
-    source_dir.mkdir()
-    source_file = source_dir / "tmp.txt"
-    with open(source_file, "w") as src_file:
-        src_file.write("Hello World")
-
-    working_dir = tmp_path / "working_dir"
-    working_dir.mkdir()
-
-    terminal = Terminal(working_dir=working_dir)
-    # Source must be a folder.
-    with pytest.raises(ValueError, match="Source .* must be a directory."):
-        terminal.copy_content(source_file)
-
-    terminal.copy_content(source_dir)
-
-    # Clean up the temporary source_dir
-    source_file.unlink()
-    source_dir.rmdir()
-
-    # Verify the content was copied correctly
-    with open(working_dir / "tmp.txt", "r") as f:
-        content = f.read()
-    assert content == "Hello World"
